make arrow data loading work with training script
