+++ scripts/train_causal.py
@@ -1,17 +1,20 @@
 import argparse
-from contextlib import nullcontext
 import os
+from contextlib import nullcontext
 from pathlib import Path
+
 import torch
 from torch import distributed as dist
 from torch import nn
 from torch.utils.data import DataLoader
 from torch.utils.data.distributed import DistributedSampler
-from fms import models
-from fms import datasets
+
+from fms import datasets, models
 from fms.models.hf.utils import to_hf_api
+from fms.training import plugins as trainplugins
+from fms.training import trainer
 from fms.utils import print0, tokenizers
-from fms.training import trainer, plugins as trainplugins
+
 
 #
 # This is a fairly minimal training/tuning script for causal language models.
@@ -145,6 +148,8 @@ if args.distributed is not None:
     dist.init_process_group(backend="nccl", rank=local_rank, world_size=world_size)
     group = dist.GroupMember.WORLD
 
+assert group.rank() == local_rank
+
 
 def get_loss_fn():
     ce_loss = torch.nn.CrossEntropyLoss()
@@ -190,7 +195,7 @@ def peft_model(model):
     from fms.models.hf.llama.modeling_llama_hf import HFAdaptedLLaMAForCausalLM
 
     model = to_hf_api(model)
-    from peft import get_peft_config, get_peft_model, LoraConfig
+    from peft import LoraConfig, get_peft_config, get_peft_model
     from peft.mapping import PeftModelForCausalLM
 
     lora_config = LoraConfig(
@@ -271,8 +276,11 @@ def main():
     )
 
     sampler = None
-    shuffle = True
-    if args.distributed == "fsdp":
+    # if the dataset is iterable, we can't shuffle it, and it should handle
+    # sharding internally
+    shuffle = not isinstance(dataset, datasets.IterableDataset)
+
+    if args.distributed == "fsdp" and not isinstance(dataset, datasets.IterableDataset):
         sampler = DistributedSampler(
             dataset, rank=local_rank, num_replicas=world_size, shuffle=True
         )
